backend/requirements.txt
fastapi
uvicorn
python-dotenv
pyjwt
supabase
python-multipart
pydantic
pydantic[email]
httpx
